+++ CHANGES.md
@@ -18,6 +18,7 @@ Translations:
 
 Build:
  - Fix issue with version name (#533)
+ - Fix rendering issue of accepted third party invitation event
 
 Changes in RiotX 0.4.0 (2019-XX-XX)
 ===================================================